requirements.txt
discord.py
google-search-results
python-dotenv
